Tinkering with the proxy code
Splits code into separate classes. Neatens up a few bits and pieces
Might want to look into configuring to use logging properly...
