+++ SpiNNaker-proxy/pom.xml
@@ -21,6 +21,9 @@ limitations under the License.
         <version>7.1.0-SNAPSHOT</version>
     </parent>
     <artifactId>SpiNNaker-proxy</artifactId>
+    <properties>
+		<mainClass>uk.ac.manchester.spinnaker.proxy.TCPProxy</mainClass>
+	</properties>
 
     <description>A simple persistent proxy server.</description>
     <name>SpiNNaker Proxy Server</name>
@@ -34,7 +37,7 @@ limitations under the License.
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>uk.ac.manchester.spinnaker.proxy.TCPProxy</mainClass>
+                            <mainClass>${mainClass}</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
