+++ SpiNNaker-proxy/src/main/java/uk/ac/manchester/spinnaker/proxy/TCPProxy.java
@@ -21,9 +21,16 @@ import java.net.Socket;
 
 /**
  * A TCP Proxy server that can re-connect if the target goes down.
+ * <p>
+ * Typically called via:
+ *
+ * <pre>java -jar spinnaker-proxy.jar LOCAL_PORT REMOTE_HOST REMOTE_PORT</pre>
+ *
+ * Writes messages to standard error describing connects and disconnects.
+ *
+ * @author Andrew Rowley
  */
 public class TCPProxy {
-
 	/** The maximum size of a read from a TCP socket. */
 	private static final int BUFFER_SIZE = 4096;
 
@@ -32,25 +39,25 @@ public class TCPProxy {
 	private final Remote remote;
 
 	TCPProxy(Socket client, String remoteHost, int remotePort) {
-		System.err.println("New connection from "
-				+ client.getRemoteSocketAddress());
+		System.err.format("New connection from %s%n",
+				client.getRemoteSocketAddress());
 		this.client = client;
 		this.remote = new Remote(remoteHost, remotePort);
 
-		Thread clientToRemote = new Thread(this::clientToRemote);
-		Thread remoteToClient = new Thread(this::remoteToClient);
+		var clientToRemote = new Thread(this::clientToRemote);
+		var remoteToClient = new Thread(this::remoteToClient);
 		clientToRemote.start();
 		remoteToClient.start();
 	}
 
 	/**
-	 * Read the client and send to the remote.  If remote write fails,
-	 * reconnect to remote and resend.  If client read fails, stop.
+	 * Read the client and send to the remote. If remote write fails, reconnect
+	 * to remote and resend. If client read fails, stop.
 	 */
 	private void clientToRemote() {
 		try {
 			remote.connect();
-			byte[] buffer = new byte[BUFFER_SIZE];
+			var buffer = new byte[BUFFER_SIZE];
 			try (var input = client.getInputStream()) {
 				while (client.isConnected()) {
 					int bytesRead = input.read(buffer);
@@ -70,13 +77,13 @@ public class TCPProxy {
 	}
 
 	/**
-	 * Read from the remote and send to the client.  If the remote read fails,
-	 * retry until working.  If the client write fails, stop.
+	 * Read from the remote and send to the client. If the remote read fails,
+	 * retry until working. If the client write fails, stop.
 	 */
 	private void remoteToClient() {
 		try {
 			remote.connect();
-			byte[] buffer = new byte[BUFFER_SIZE];
+			var buffer = new byte[BUFFER_SIZE];
 			try (var output = client.getOutputStream()) {
 				while (client.isConnected()) {
 					int bytesRead = remote.read(buffer);
@@ -102,8 +109,7 @@ public class TCPProxy {
 		if (client == null) {
 			return;
 		}
-		System.err.println(
-				"Client " + client.getRemoteSocketAddress() + " left");
+		System.err.format("Client %s left%n", client.getRemoteSocketAddress());
 		try {
 			client.close();
 		} catch (IOException e) {
@@ -114,14 +120,17 @@ public class TCPProxy {
 
 	/**
 	 * The main method.
-	 * @param args args[0]: The local port to listen on
-	 *             args[1]: The remote host to proxy
-	 *             args[2]: The remote port to proxy
-	 * @throws IOException If we can't start the server.
+	 *
+	 * @param args
+	 *            {@code args[0]}: The local port to listen on <br>
+	 *            {@code args[1]}: The remote host to proxy <br>
+	 *            {@code args[2]}: The remote port to proxy
+	 * @throws IOException
+	 *             If we can't start the server.
 	 */
 	public static void main(String[] args) throws IOException {
 		int localPort = Integer.parseInt(args[0]);
-		String remoteHost = args[1];
+		var remoteHost = args[1];
 		int remotePort = Integer.parseInt(args[2]);
 
 		try (var server = new ServerSocket(localPort)) {
@@ -129,148 +138,6 @@ public class TCPProxy {
 				var client = server.accept();
 				new TCPProxy(client, remoteHost, remotePort);
 			}
-		}
-	}
-}
-
-/**
- * A persistent remote connection.
- */
-final class Remote {
-
-	/** How long to wait between retries of the connection. */
-	private static final int RETRY_MS = 5000;
-
-	/** How long to wait between writes on a connection. */
-	private static final int WRITE_RETRY_MS = 1000;
-
-	private final String remoteHost;
-
-	private final int remotePort;
-
-	private Socket remote = null;
-
-	private boolean closed = false;
-
-	/**
-	 * Create a remote.
-	 *
-	 * @param remoteHost The host to connect to.
-	 * @param remotePort The port to connect to.
-	 */
-	Remote(String remoteHost, int remotePort) {
-		this.remoteHost = remoteHost;
-		this.remotePort = remotePort;
-	}
-
-	/**
-	 * Connect to the remote site and don't stop until connected.
-	 *
-	 * @throws InterruptedException If interrupted.
-	 */
-	synchronized void connect() throws InterruptedException {
-		if (remote != null || closed) {
-			return;
-		}
-		System.err.println("Connecting to " + remoteHost + ":" + remotePort);
-		while (true) {
-			try {
-				remote = new Socket(remoteHost, remotePort);
-				System.err.println("Connected to "
-						+ remote.getRemoteSocketAddress());
-				return;
-			} catch (IOException e) {
-				Thread.sleep(RETRY_MS);
-			}
-		}
-	}
-
-	/**
-	 * Write to the remote.  Retry until successful, or interrupted.
-	 *
-	 * @param buffer The buffer to write.
-	 * @param offset The offset of the buffer to start from.
-	 * @param length The length of the buffer to write, starting at the offset.
-	 * @throws InterruptedException If interrupted.
-	 */
-	void write(byte[] buffer, int offset, int length)
-			throws InterruptedException {
-		while (true) {
-			while (remote == null) {
-				Thread.sleep(WRITE_RETRY_MS);
-			}
-			try {
-				var remoteOutput = remote.getOutputStream();
-				remoteOutput.write(buffer, 0, length);
-				return;
-			} catch (IOException e) {
-				if (!closed) {
-					closeConnections();
-					connect();
-				}
-			}
-		}
-	}
-
-	/**
-	 * Read from the remote.  Retry until successful, or interrupted.
-	 *
-	 * @param buffer The buffer to read into.
-	 * @return The bytes written.
-	 * @throws InterruptedException If interrupted.
-	 */
-	int read(byte[] buffer) throws InterruptedException {
-		while (true) {
-			if (remote == null) {
-				return -1;
-			}
-			try {
-				var remoteInput = remote.getInputStream();
-				int bytesRead = remoteInput.read(buffer);
-				if (bytesRead == -1) {
-					throw new IOException();
-				}
-				return bytesRead;
-			} catch (IOException e) {
-				if (!closed) {
-					closeConnections();
-					connect();
-				}
-			}
-		}
-	}
-
-	/**
-	 * Close the connection never to be reopened.
-	 */
-	void close() {
-		closed = true;
-		closeConnections();
-	}
-
-	/**
-	 * Close the connections.
-	 */
-	synchronized void closeConnections() {
-		if (remote == null) {
-			return;
-		}
-		forceClose(remote);
-		remote = null;
-	}
-
-	/**
-	 * Force closure and ignore any exceptions.
-	 * @param closable The thing to close.
-	 */
-	private void forceClose(AutoCloseable closable) {
-		if (closable == null) {
-			return;
-		}
-		try {
-			closable.close();
-		} catch (Exception e) {
-			// Ignore the exception
 		}
 	}
 }
